saqlandi

+++ app/src/main/java/uz/orifjon/educationsysteminandroid/database/MySqliteHelper.kt
@@ -336,7 +336,7 @@ class MySqliteHelper(context: Context) : SQLiteOpenHelper(
                 val lastname = cursor.getString(2)
                 val patron = cursor.getString(3)
                 val registerDate = cursor.getString(4)
-                val groupId = cursor.getInt(5)
+                val groupId = cursor.getLong(5)
                 val student = Student(
                     id = id1,
                     firstname = firstname,
